pyproject.toml
[project]
name = "realtime-whispyannote"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.11"
dependencies = [
    "openvino ==2024.1.0",
    "optimum-intel ==1.22.0",
    "pyannote-audio ==3.3.2",
    "pyaudio ==0.2.14",
    "pyside6 ==6.9.0",
    "torch ==2.6.0",
    "torchaudio ==2.6.0",
    "transformers ==4.36.2",
]

[project.scripts]
whispyannote = "whispyannote.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/whispyannote"]
